requirements.txt
numpy==1.24.3
pandas==2.2.3
plotly==5.9.0
protobuf
python-dotenv==1.0.1
streamlit==1.34.0
langchain
langchain-ollama
